update (SMAPI 3.18.6, SDV 1.5.6)

+++ source/Stardew Druid/StardewDruid/Monster/BossBat.cs
@@ -11,10 +11,8 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewValley;
-using StardewValley.Monsters;
 using System;
 using System.Collections.Generic;
-using System.Reflection.Metadata.Ecma335;
 
 namespace StardewDruid.Monster
 {
@@ -35,12 +33,12 @@ namespace StardewDruid.Monster
         public Dictionary<int, Vector2> hatOffsets;
 
         public BossBat(Vector2 vector, int combatModifier)
-            : base(vector * 64,200)
+            : base(vector * 64, 200)
         {
 
             moveTowardPlayerThreshold.Value = 99;
 
-            Health = (int)(combatModifier * 6);
+            Health = combatModifier * 6;
 
             MaxHealth = Health;
 
@@ -82,7 +80,7 @@ namespace StardewDruid.Monster
         {
 
             if (posturing) { return; }
-            
+
             base.behaviorAtGameTick(time);
 
         }
@@ -93,7 +91,7 @@ namespace StardewDruid.Monster
             if (posturing) { return; }
 
             base.defaultMovementBehavior(time);
-            
+
         }
 
         protected override void localDeathAnimation()
@@ -101,7 +99,7 @@ namespace StardewDruid.Monster
             ModUtility.AnimateDeathSpray(base.currentLocation, base.Position, Color.MediumPurple);
         }
 
-        public override int takeDamage(int damage, int xTrajectory, int yTrajectory, bool isBomb, double addedPrecision, Farmer who) 
+        public override int takeDamage(int damage, int xTrajectory, int yTrajectory, bool isBomb, double addedPrecision, Farmer who)
         {
 
             if (posturing) { return 0; }
@@ -114,7 +112,7 @@ namespace StardewDruid.Monster
 
             }
 
-            return base.takeDamage(damage, xTrajectory,yTrajectory,isBomb,addedPrecision,who);
+            return base.takeDamage(damage, xTrajectory, yTrajectory, isBomb, addedPrecision, who);
 
         }
 
@@ -126,27 +124,27 @@ namespace StardewDruid.Monster
             }
 
             b.Draw(
-                Sprite.Texture, 
-                getLocalPosition(Game1.viewport)+ new Vector2(32f, 32f), 
-                Sprite.SourceRect, 
-                (shakeTimer > 0) ? Color.Red : Color.White, 
-                0f, 
-                new Vector2(8f, 16f), 
-                Math.Max(0.2f, scale) * 6f, 
+                Sprite.Texture,
+                getLocalPosition(Game1.viewport) + new Vector2(32f, 32f),
+                Sprite.SourceRect,
+                (shakeTimer > 0) ? Color.Red : Color.White,
+                0f,
+                new Vector2(8f, 16f),
+                Math.Max(0.2f, scale) * 6f,
                 flip ? SpriteEffects.FlipHorizontally : SpriteEffects.None,
                 Math.Max(0f, drawOnTop ? 0.991f : getStandingY() / 10000f)
             );
-            
+
             b.Draw(
-                Game1.shadowTexture, 
-                getLocalPosition(Game1.viewport)+ new Vector2(32f, 64f), 
-                Game1.shadowTexture.Bounds, 
-                Color.White, 
-                0f, 
-                new Vector2(Game1.shadowTexture.Bounds.Center.X, Game1.shadowTexture.Bounds.Center.Y), 
-                6f, 
-                SpriteEffects.None, 
-                base.wildernessFarmMonster ? 0.0001f : ((float)(getStandingY() - 1) / 10000f)
+                Game1.shadowTexture,
+                getLocalPosition(Game1.viewport) + new Vector2(32f, 64f),
+                Game1.shadowTexture.Bounds,
+                Color.White,
+                0f,
+                new Vector2(Game1.shadowTexture.Bounds.Center.X, Game1.shadowTexture.Bounds.Center.Y),
+                6f,
+                SpriteEffects.None,
+                base.wildernessFarmMonster ? 0.0001f : ((getStandingY() - 1) / 10000f)
              );
 
             int frameOffset = Sprite.currentFrame % 4;
