+++ source/~drbirbdev/GameboyArcade/CoreBoy/sound/NullSoundOutput.cs
@@ -20,7 +20,7 @@ namespace CoreBoy.sound
         {
         }
 
-        public void Play(int left, int right)
+        public void Play(byte left, byte right)
         {
         }
     }
